Show configurable test states, with extra formatting (#1, #2)
This commit includes a little refactoring, so it is not as tidy as
I normally would like.

Essentially, this commit allows the extension to show tests which have
passed, errored, or been skipped, alongside those which have failed.

To make this work, several changes to how I format test diagnostics were
made to genericize them. Therefore, **test diagnostics will read
differently** and include the status and the label of the test.

Note that this implies a possibly considerable regression: Right now,
I've decided only to look at the first test event decoration when
deciding which diagnostics to show. It is my intention to address this
in the future if need be. But I am not sure if or how many test adapters
actually make use of the ability to set multiple decorations.

I have added clear "TODO" comments to represent decisions I made which
can be improved upon. I intend to open issues for these.

See #1, #2.