Note that I intend to allow customizing diagnostics
This is a TODO for me, but I also want to make it obvious that it's
going to happen. Or at least that I hope for it to happen.

Note also that I've finished adding the ability to customize which
diagnostics are shown, and will be releasing that very soon.